/ allowing decimal numbers in throttle option
--HG--
extra : convert_revision : svn%3A05628a4a-d5d7-11dd-a2b5-e5f827957e07/trunk%4085

+++ log4tailer.py
@@ -70,7 +70,7 @@ def main():
     if options.pause:
         pause = int(options.pause)
     if options.throttle:
-        throttle = int(options.throttle)
+        throttle = float(options.throttle)
     if options.silence:
         # silence mode enables sendSmtp Action.
         # There could be other actions like Post to Web Page
